requirements.txt
absl-py==0.13.0
adamp==0.3.0
albumentations==1.0.3
antlr4-python3-runtime==4.8
appnope==0.1.2
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.6.0
black==21.9b0
bleach==3.3.1
blis==0.7.4
cachetools==4.2.2
catalogue==2.0.6
certifi==2020.4.5.1
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
clang==5.0
click==7.1.2
cmake==3.21.2
colab-ssh==0.3.17
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.5
Cython==0.29.24
dataclasses==0.6
datasets==1.11.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.4
Distance==0.1.3
dlib==19.19.0
docker-pycreds==0.4.0
dotmap==1.3.14
easydict==1.9
efficientnet-pytorch==0.7.1
einops==0.3.0
emoji==1.5.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0-py3-none-any.whl
entrypoints==0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
fairseq==0.10.2
faiss-cpu==1.7.1.post2
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.0.12
flatbuffers==1.12
fsspec==2021.8.1
future==0.18.2
g2p-en==2.1.0
gast==0.4.0
gdown==3.13.0
gitdb==4.0.7
GitPython==3.1.24
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
huggingface-hub==0.0.16
hydra-core==1.1.1
idna==2.10
imageio==2.9.0
importlib-resources==5.2.2
imutils==0.5.3
inflect==5.3.0
ipykernel==6.0.3
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
jieba3k==0.35.1
Jinja2==3.0.1
joblib==1.0.1
JPype1==1.3.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras==2.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
konlpy==0.5.2
kss==3.2.0
lxml==4.6.3
marisa-trie==0.7.7
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mecab-python===0.996-ko-0.9.2
mistune==0.8.4
mock==4.0.3
mpmath==1.2.1
multiprocess==0.70.12.2
murmurhash==1.0.5
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.2
newspaper3k==0.2.8
nltk==3.6.2
notebook==6.4.0
numpy==1.19.2
oauthlib==3.1.1
omegaconf==2.1.1
opencv-python==4.2.0.34
opencv-python-headless==4.5.3.56
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.9.0
pathtools==0.1.2
pathy==0.6.0
petitions-dataset==1.0.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
platformdirs==2.3.0
pororo==0.4.2
portalocker==2.3.2
preshed==3.0.5
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
py-hanspell @ git+https://github.com/ssut/py-hanspell.git@8e993cf46f97f9d665c15633a0fc78ac1b727713
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
Pygments==2.9.0
pykospacing @ git+https://github.com/haven-jeon/PyKoSpacing.git@948e8988450abf631df515ff4151b6ebee864ba3
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
regex==2021.8.28
requests==2.24.0
requests-file==1.5.1
requests-oauthlib==1.3.0
retina-face==0.0.5
rsa==4.7.2
sacrebleu==2.0.0
sacremoses==0.0.45
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
selenium==3.141.0
Send2Trash==1.7.1
sentence-transformers==2.0.0
sentencepiece==0.1.96
sentry-sdk==1.4.2
sgmllib3k==1.0.0
shortuuid==1.0.1
six==1.15.0
smart-open==5.2.1
smmap==4.0.0
soynlp==0.0.493
spacy==3.1.2
spacy-legacy==3.0.8
srsly==2.4.1
subprocess32==3.5.4
sympy==1.8
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.1
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
thinc==8.0.9
threadpoolctl==2.2.0
tifffile==2021.8.8
timm==0.4.12
tinysegmenter==0.3
tldextract==3.1.2
tokenizers==0.10.3
tomli==1.2.1
torch==1.6.0
torchsummary==1.5.1
torchtext==0.10.0
torchvision==0.7.0
tornado==6.1
tqdm==4.46.0
traitlets==5.0.5
transformers==4.10.0
tweepy==3.10.0
typer==0.3.2
typing-extensions==3.10.0.2
urllib3==1.25.11
wandb==0.12.3
wasabi==0.8.2
wcwidth==0.2.5
webdriver-manager==3.4.2
webencodings==0.5.1
Werkzeug==2.0.1
wget==3.2
Whoosh==2.7.4
widgetsnbextension==3.5.1
wincertstore==0.2
word2word==1.0.0
wrapt==1.12.1
xxhash==2.0.2
yaspin==2.1.0
zipp==3.6.0
